+++ src/views/MainWindow/MainWindow.tsx
@@ -126,10 +126,6 @@ function MainWindow(props: MainWindowProps) {
   }, []);
 
   useEffect(() => {
-    window.api.menu.setLockScreenActive(false);
-  }, []);
-
-  useEffect(() => {
     window.api.storage.getNotes()
       .then((notes: NoteType[]) => {
         setNotes(sortNotes(notes, currentNotesSortOrder.current));
@@ -187,7 +183,9 @@ function MainWindow(props: MainWindowProps) {
 
   useEffect(() => {
     const offMenuNewNote = window.api.menu.onMenuNewNote(handleAddNote);
-    const offMenuLockNotes = window.api.menu.onMenuLockNotes(() => navigate(AppView.lock));
+    const offMenuLockNotes = window.api.menu.onMenuLockNotes(() => {
+      window.api.security.lock();
+    });
     const offMenuShowWelcome = window.api.menu.onMenuShowWelcome(() => navigate(AppView.welcome));
     const offMenuSelectNote = window.api.menu.onMenuSelectNote(() => setIsSelectionMode(true));
     const offMenuSelectAllNotes = window.api.menu.onMenuSelectAllNotes(handleToggleSelectAllNotes);